+++ front/src/api/auth.js
@@ -1,0 +1,18 @@
+import { instance } from "./index";
+
+function loginUser(userData) {
+  console.log(userData);
+  return instance.post("login", userData);
+}
+function signupUser(userData) {
+  console.log(userData);
+  return instance.post(instance.url, userData);
+}
+function certUserEmail(Email, certNum) {
+  return instance.post(instance.url, Email, certNum);
+}
+function userIdCheck(userId) {
+  return instance.post(instance.url, userId);
+}
+
+export { loginUser, signupUser, certUserEmail, userIdCheck };
