+++ front/src/api/hotel.js
@@ -1,0 +1,7 @@
+import { instance } from "./index";
+
+function hotelList(location) {
+  return instance.post("/hotel", location);
+}
+
+export { hotelList };
